+++ lib/data/database/tables/payment.dart
@@ -4,9 +4,10 @@ import 'package:trackpense/data/constants/kjp_miscellaneous.dart';
 class Payment extends Table {
   TextColumn get uuid => text().clientDefault(() => KjpMiscellaneous.uuid.v4())();
   DateTimeColumn get date => dateTime()();
-  TextColumn get description => text().withLength(min: 1, max: 255)();
+  TextColumn get description => text().withLength(min: 1, max: 200)();
   RealColumn get amount => real()();
   BoolColumn get isExpense => boolean()();
+  TextColumn get notes => text().nullable()();
 
   @override
   Set<Column> get primaryKey => {uuid};