+++ CHANGELOG.md
@@ -2,3 +2,4 @@
  - Income/Expense is now selected with a switch instead of radio buttons.
  - Date and time can now be set to any day within the last 7 days.
  - Improved styling on cards
+ - Added new field to include any required notes